ci: add Makefile target and CI check for collectstatic (#65)
* ci(static): Add CI check for up-to-date static files

This commit introduces a new check in the CI pipeline to ensure that Django's collected static files are always up-to-date and committed to the repository.

A new `lint` target has been added to the `dashboard/Makefile`. This target first runs `make collectstatic` and then uses `git diff` to verify that there are no uncommitted changes in the `staticfiles/` directory. If changes are detected, the build will fail.

The GitHub Actions workflow has been updated to execute this new `make lint` command, automating the verification process. This prevents developers from forgetting to run `collectstatic` after modifying assets, which could lead to deploying outdated static files.

* add Makefile help for collectstatic

* rename lint to collectstatic-check

* add README section about static files

* rename Tests workflow to Checks

* change one more instance of lint

* adjust new README section

---------

Co-authored-by: David Wilding <david.wilding@canonical.com>