Improved performance of Project list view (main dashboard view) (#77)
* Fixes incorrect status level

When an objective's commitments don't include a level, the status
result could be wrong, reporting the missing level as achieved.

This fixes that by only checking for levels that are actually
part of the objective's commitments.

Includes a test.

* Improved performance of Project list view (main dashboard view)

For example:

Before: 1799 SQL queries in 80ms, rendered in 1471ms
Now   : 1352 SQL queries in 55ms, rendered in 1015ms

Still too slow, but a good improvement.

* bump rock version

---------

Co-authored-by: David Wilding <david.wilding@canonical.com>