Unify ProjectObjectiveCondition status into a single field (#83)
* Added missing migrations

* Added status field to ProjectObjectiveCondition

* Added data migration for unified status field

* Completed migration to unified ProjectObjectiveCondition.status field

The data migration removes the three old done/not_applicable/candidate
fields. Now the status field contains one of four values:

* "" (default)
* "DO": done
* "NA": not applicable
* "CA": candidate

provided as Choices field.

This simplifies management, and makes it possible to reduce the number
of URL patterns and view functions required from three to one.

The new URL pattern for managing a ProjectObjectiveCondition.status is

  /action_toggle_condition/<id>?status=<status>&target=<target>

where <id> is the ID of the ProjectObjectiveCondition, <status> is
its current status (one of the four values described above) and <target>
is the interface element interacted with.

The htmx delays of .5s in partial_condition.html have been removed.
The action_toggle_condition view returns the ProjectObjectiveCondition
object directly to the template partial, so there is nothing to wait
for.

Previously, view toggled database values based on what the user's actions.
This meant it was possible for the user to see a value in the interface
that was out of date (having already been updated in another session).
Their attempt to correct it would toggle the database value, meaning
that they could inadvertently change correct data.

The view now evaluates the new status value for the ProjectObjectiveCondition
based not on its database value, but the value as shown in the template
(and therefore the user's interface). The value saved will always be the
value the user expects.

Also removed the driver user from the initial_data.yaml - not really
necessary, and it introduced fragility.

* update static files

* bump rock version

---------

Co-authored-by: David Wilding <david.wilding@canonical.com>